@@ -7,11 +7,11 @@ usually due to method overloading or diamond inheritance (in interfaces).*/
 public class AmbiguityDemo {
 
 	public static void sum(int a, long b) {
-		System.out.println("Sum of int and long value :" + (a + b));
+        System.out.println("Sum of int and long value :" + (a + b));
 	}
 
 	public static void sum(long a, int b) {
-		System.out.println("Sum of long and int value :" + (a + b));
+        System.out.println("Sum of long and int value :" + (a + b));
 	}
 	
 	public static void main(String[] args) {
@@ -19,6 +19,6 @@ public class AmbiguityDemo {
 		//AmbiguityDemo.sum(10, 20);
 		
 		//Solution.
-		AmbiguityDemo.sum(10, 20l);
+		AmbiguityDemo.sum(10, 20L);
 	}
 }
