@@ -7,13 +7,14 @@ public class ExecutorExample {
     public static void main(String[] args) {
 
         // Create a thread pool with 3 threads
-        ExecutorService executor = Executors.newFixedThreadPool(3);
-        Runnable task = () -> {
-            System.out.println("Task executed by: " + Thread.currentThread().getName());
-        };
-        for (int i = 0; i < 5; i++) {
-            executor.submit(task);
+        try (ExecutorService executor = Executors.newFixedThreadPool(3)) {
+            Runnable task = () -> System.out.println("Task executed by: " +
+                    Thread.currentThread()
+                    .getName());
+            for (int i = 0; i < 5; i++) {
+                executor.submit(task);
+            }
+            executor.shutdown(); // Initiates an orderly shutdown
         }
-        executor.shutdown(); // Initiates an orderly shutdown
     }
 }
