@@ -3,7 +3,8 @@ package org.myapp.interview.strings;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-public class Program_1 {
+public class Program1 {
+
     public static void main(String[] args) {
 
         String str = "swiss";
