@@ -4,7 +4,7 @@ import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.stream.Collectors;
 
-public class Program_4 {
+public class Program4 {
 
     public static void main(String[] args) {
 
