@@ -9,9 +9,9 @@ public class ReverseStringUsingRecursion {
 
     public static String reverse(String str) {
         // Base case: if the string is empty or has only one character
-        if (str == null || str.isBlank() || str.length() <= 1) {
+        if (str == null || str.isBlank() || str.length() == 1)
             return str;
-        }
+
         // Recursive case: reverse the rest of the string and add the first character at the end
         return reverse(str.substring(1)) + str.charAt(0);
     }
@@ -19,7 +19,7 @@ public class ReverseStringUsingRecursion {
     public static void main(String[] args) {
         String str = "vivek";
         String reversed = reverse(str);
-        log.info("Reversed string: " + reversed);
+        log.info("Reversed String: {}", reversed);
     }
 }
 
